maake detail room to be one file
